Fix checking against outdated window bounds
Fixes #17.  When insertmode checked t > end, it was checking against the window
boundaries _before_ the latest insertion.  When a small amount of text was on
screen, end would point to the end of the buffer, rather than the end of what
fits on screen (since the buffer is small enough to fit on screen entirely).
So after inserting a character and advancing t, t would be > end, and a scroll
would be triggered.  This scroll would put all of the text off the top of the
screen.  Fixing the check to happen after drawing makes it use updated
boundaries ensures that a scroll only happens if the insert target is actually
off screen.