Only trim last character off of winbounds if it is a newline
winbounds trims the last character to avoid drawing an extra newline
for the last on screen line.  With small buffers that don't have a
trailing newline, this would cause incorrect behavior.  The last
character would be cut off from view and inserting at the last
character would always cause the text to scroll out of view.