Remove problematic winbounds adjustment code
This code has caused a lot of out-of-bounds accesses in the past, and is
currently causing assertions for empty files.

I think this was originally added because the window boundaries included
an unneeded newline.  Back before curses was used for drawing, the extra
newline would have pushed things off of the screen.  I don't think it's
needed any more.  In testing there didn't seem to be any negative
effects.  The same amount of text was drawn on screen, nothing was
pushed off the top or anything.  Targeting code seemed fine.