Fix assertion firing when moving cursor to beginning of line
When we're inserting at the beginning of a line, the skipscreenlines call
will return exactly the target location.  Since we don't have any further
offset to add we shouldn't enter the loop at all.  The wrong comparison
when finding the row caused the loop to be entered and the assertion
happened to catch it.