Reduce flickering cursor when typing
I noticed that the cursor flickers when typing.  This is because the
buffer is drawn and presented, which moves the cursor to the bottom
right, and then the cursor is moved back into the correct position.
If we reorder things so that the move happens before present() is
called, then we lose the flicker.