Pipe selected text through a sample command
I added the ability to fork and communicate with the subprocess.
For testing purposes I don't query for a command yet.

While testing this I realized that it will be very useful if we collect
stderr from the child as well, so that we can display it when there's
a problem.