Fix incorrect line targeting when lines wrap
The problem here turned out to be that we were compensating for long
lines twice.  The functions that were using the offset from huntline
were using skipscreenlines to convert that to a coordinate in the buffer.
But the offsets returned from huntline are actually buffer line offsets,
not screen lines.  So when a wrapped line occurs we would actually skip
an extra line when performing that conversion.