Use a pointer to keep the scroll location
`refresh_bounds` was horribly slow, because every time it was called it
would scroll through the file from the beginning to skip N scroll lines.
Now we keep a pointer to the position in the file that we are looking at,
so that it doesn't need to be recalculated each time.