Only recalculate bounds when needed
I switched a few loops to memchr which seems to provide a small boost.
The bigger gain though comes from updating the window bounds when
the buffer changes or the window is scrolled.  This cuts down on the
expensive linear search for the end of the window, which was being run
too often.

The downside is that it's now up to app code to refresh the bounds when
the buffer changes, instead of automatically refreshing it whenever the
bounds are needed.