Track changes to the buffer for undo
In order to undo actions, we have to track them.  Every change to
the buffer boils down to inserts and deletes, and they are the
inverse of each other.  By tracking the positions of inserted text,
we can perform a delete operation to undo the insertion.  By tracking
the position and contents of deleted text, we can perform an insertion
to undo the deletion.

Adding the undo module to store the data structures for tracking
these commands.