Add some basic coloring
The colors seem to be more reliable than DIM and STANDOUT.  At least for me
tmux doesn't cooperate with standout.  I know I had colors and then removed
them because they don't seem to fit with the minimal feel of lwe, but it
really is useful to have targets highlighted consistently, so here they
are again.